refactor: wip pages effectifs + enquete-sifa + harmonisation style
