+++ ui/pages/effectifs/aide-configuration-erp.tsx
@@ -5,25 +5,26 @@ import Head from "next/head";
 import Page from "@/components/Page/Page";
 import withAuth from "@/components/withAuth";
 import { getAuthServerSideProps } from "@/common/SSR/getAuthServerSideProps";
+import ConfigurationAPI from "@/modules/mon-espace/effectifs/ConfigurationAPI";
 
 export const getServerSideProps = async (context) => ({ props: { ...(await getAuthServerSideProps(context)) } });
 
-const PageImportEffectifsDeMonOrganisme = () => {
-  const title = "Import";
-
-  // FIXME ajouter l'organisme
+const PageAideConfigurationErp = () => {
+  const title = "Configuration ERP";
+  // apparemment pas besoin de l'organisme ici
   return (
     <Page>
       <Head>
-        <title>Mes effectifs - {title}</title>
+        <title>{title}</title>
+        <link rel="icon" href="/favicon.ico" />
       </Head>
       <Box w="100%" pt={[4, 6]} px={[1, 1, 2, 4]} mb={16}>
         <Container maxW="xl" px={0}>
-          <Box mt={4}>{/* <Televersements organisme={myOrganisme} /> */}</Box>
+          <ConfigurationAPI />
         </Container>
       </Box>
     </Page>
   );
 };
 
-export default withAuth(PageImportEffectifsDeMonOrganisme);
+export default withAuth(PageAideConfigurationErp);
