+++ ui/pages/effectifs/televersement.tsx
@@ -5,27 +5,26 @@ import Head from "next/head";
 import Page from "@/components/Page/Page";
 import withAuth from "@/components/withAuth";
 import { getAuthServerSideProps } from "@/common/SSR/getAuthServerSideProps";
-import ConfigurationAPI from "@/modules/mon-espace/effectifs/ConfigurationAPI";
+import { useEffectifsOrganismeOrganisation } from "@/modules/mon-espace/effectifs/useEffectifsOrganisme";
+import Televersements from "@/modules/mon-espace/effectifs/Televersements";
 
 export const getServerSideProps = async (context) => ({ props: { ...(await getAuthServerSideProps(context)) } });
 
-const PageAideConfigurationErp = () => {
-  const title = "Configuration ERP";
+const PageImportEffectifsDeMonOrganisme = () => {
+  const title = "Import";
+
+  const { organisme } = useEffectifsOrganismeOrganisation();
+
   return (
     <Page>
       <Head>
-        <title>{title}</title>
-        <link rel="icon" href="/favicon.ico" />
+        <title>Mes effectifs - {title}</title>
       </Head>
       <Box w="100%" pt={[4, 6]} px={[1, 1, 2, 4]} mb={16}>
-        <Container maxW="xl" px={0}>
-          <Box mt={4}>
-            <ConfigurationAPI />
-          </Box>
-        </Container>
+        <Container maxW="xl">{organisme && <Televersements organisme={organisme} />}</Container>
       </Box>
     </Page>
   );
 };
 
-export default withAuth(PageAideConfigurationErp);
+export default withAuth(PageImportEffectifsDeMonOrganisme);
